Swiping Feature Update
Updated the swiping so that its based not on velocity but on the position of the card. The card can be released based on its position on the screen and will give feedback based on where it is (red for rejected, green for favorited).
----------------------------
Card.js
Checks for the position of the card because it will increase the efficiency of the swipe.

Card.module.css
Implemented css so that the card will give live feedback.