Login/Signin Feedback To Wrong Input/ Modal View for Home Page / Feedback on new swiping
Changes to Login/Sign-in where the forms give live feedback to when the user types in an invalid form input.
Changes to Signup.js where the user was able to insepct element and remove the key from the console and be able to sign up without typing a user name using Formik package. Now there is more security on the signup process.
Changes to Login.js where feedback is given if the wrong username/password is typed in. Before this was not implemented.
Changes to SwipeableCard.js where the live feeedback is given on a swipe, and when the user clicks the bottom of the card, the modal view of the place pops up/
Changes to corresponding CSS files for styling purposes, such as in Login.module.css where the input for the password does not overlap the reveal icon on the password form.
Changes to CardDeck.js, Card.js to incorporate the modal view.
Added Text.js for favorite details.

Apologies for the big update, should have been separated into separate PULL REQUESTS.